requirements.txt
openai
imapclient
python-dotenv
yagmail
PyYAML